add home page and route it
