+++ src/index.tsx
@@ -6,6 +6,7 @@ import reportWebVitals from "./reportWebVitals"
 import { ColorModeScript } from "@chakra-ui/react"
 import { Provider } from "react-redux"
 import { store } from "./services/store"
+import { BrowserRouter } from "react-router-dom"
 
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
 root.render(
@@ -13,7 +14,9 @@ root.render(
     <ColorModeScript initialColorMode={theme.config.initialColorMode} />
     <React.StrictMode>
       <Provider store={store}>
-        <App />
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
       </Provider>
     </React.StrictMode>
   </>
